+++ packages/fetchye/__tests__/makeServerFetchye.spec.js
@@ -29,15 +29,8 @@ const defaultPayload = {
   }),
   ok: true,
   status: 200,
-  json: async () => ({
+  text: async () => JSON.stringify({
     fakeData: true,
-  }),
-  clone: () => ({
-    headers: new global.Headers({
-      'Content-Type': 'application/json',
-    }),
-    ok: true,
-    status: 200,
   }),
 };
 
@@ -120,7 +113,7 @@ describe('makeServerFetchye', () => {
       }
     `);
   });
-  it('should return error in error state', async () => {
+  it('should return null in the error state', async () => {
     const fetchClient = jest.fn(async () => {
       throw new Error('fake error');
     });
@@ -140,7 +133,7 @@ describe('makeServerFetchye', () => {
     expect(fetchyeRes).toMatchInlineSnapshot(`
       Object {
         "data": null,
-        "error": "Error: fake error",
+        "error": null,
       }
     `);
   });